# apps/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0

servers:
  - url: http://localhost:3000/api
    description: Desarrollo

tags:
  - name: Eventos
    description: Endpoints de eventos
  - name: Ausencias
    description: Endpoints de ausencias
  - name: Noticias
    description: Endpoints de noticias
  - name: Estadisticas
    description: Endpoints de estadisticas

components:
  schemas:
    Meta:
      type: object
      properties:
        total:
          type: integer
          example: 47
        limit:
          type: integer
          nullable: true
          example: 10
        offset:
          type: integer
          example: 0

    Evento:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Conferencia de tecnologia
        descripcion:
          type: string
          example: Evento sobre las ultimas tendencias
        fecha:
          type: string
          format: date-time
          example: 2026-05-01T10:00:00.000Z
        creado:
          type: string
          format: date-time
          example: 2026-04-01T08:00:00.000Z

    Ausencia:
      type: object
      description: Ausencia retornada en GET (incluye docente anidado)
      properties:
        id:
          type: integer
          example: 1
        materia:
          type: string
          example: Matematica
        creado:
          type: string
          format: date-time
          example: 2026-04-11T08:00:00.000Z
        docente:
          type: object
          properties:
            name:
              type: string
              example: Juan Perez

    AusenciaCreada:
      type: object
      description: Ausencia retornada tras POST (registro plano, sin docente anidado)
      properties:
        id:
          type: integer
          example: 1
        materia:
          type: string
          example: Matematica
        creado:
          type: string
          format: date-time
          example: 2026-04-11T08:00:00.000Z
        docenteId:
          type: string
          example: clxk2j3r40000abc123
        publicadorId:
          type: string
          example: clxk2j3r40000xyz456

    Recurso:
      type: object
      properties:
        id:
          type: integer
          example: 1
        resource:
          type: string
          example: 11-4-2026_10-30-00-123-ab3f.jpg
        fecha:
          type: string
          format: date-time
          example: 2026-04-11T10:30:00.000Z
        es_noticia:
          type: boolean
          example: true
        noticiaId:
          type: integer
          example: 1
        userId:
          type: string
          example: clxk2j3r40000abc123

    Noticia:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Nueva actualizacion del sistema
        descripcion:
          type: string
          example: Se implementaron mejoras en el rendimiento
        publicado:
          type: string
          format: date-time
          example: 2026-04-11T08:00:00.000Z
        recursos:
          type: array
          items:
            $ref: "#/components/schemas/Recurso"

    EstadisticaItem:
      type: object
      description: Objeto con una sola clave dinamica y su conteo
      additionalProperties:
        type: integer
      example:
        eventos: 5

  parameters:
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
      required: false
      description: Cantidad maxima de resultados a retornar. Si no se provee, se retornan todos los resultados sin limite.
    OffsetParam:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
      required: false
      description: Cantidad de resultados a saltar. Por defecto 0.

  responses:
    NoAutorizado:
      description: Sin sesion activa
      content:
        application/json:
          schema:
            type: object
            properties:
              mensaje:
                type: string
                example: No autorizado

    SinPermisos:
      description: Sin permisos suficientes
      content:
        application/json:
          schema:
            type: object
            properties:
              mensaje:
                type: string
                example: "Sin permisos para crear en eventos"

    ErrorServidor:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            type: object
            properties:
              mensaje:
                type: string
                example: Error en el servidor

paths:
  /eventos:
    get:
      tags: [Eventos]
      summary: Obtener eventos proximos
      description: Retorna todos los eventos cuya fecha no haya vencido aun, ordenados por fecha ascendente.
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: OK o sin resultados
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: OK
                  meta:
                    $ref: "#/components/schemas/Meta"
                  eventos:
                    type: array
                    items:
                      $ref: "#/components/schemas/Evento"
              examples:
                con_resultados:
                  summary: Con eventos
                  value:
                    mensaje: OK
                    meta:
                      total: 5
                      limit: 10
                      offset: 0
                    eventos:
                      - id: 1
                        nombre: Conferencia de tecnologia
                        descripcion: Evento sobre las ultimas tendencias
                        fecha: "2026-05-01T10:00:00.000Z"
                        creado: "2026-04-01T08:00:00.000Z"
                sin_resultados:
                  summary: Sin eventos
                  value:
                    mensaje: No hay eventos disponibles
                    meta:
                      total: 0
                      limit: 10
                      offset: 0
                    eventos: []
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Error en el servidor
                  meta:
                    $ref: "#/components/schemas/Meta"
                  eventos:
                    nullable: true
                    example: null
              example:
                mensaje: Error en el servidor
                meta:
                  total: 0
                  limit: null
                  offset: 0
                eventos: null

    post:
      tags: [Eventos]
      summary: Crear evento
      description: Middleware `requierePermiso("eventos", "crear")`. Requiere sesion activa.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, descripcion, fecha]
              properties:
                nombre:
                  type: string
                  example: Conferencia de tecnologia
                descripcion:
                  type: string
                  example: Evento sobre las ultimas tendencias
                fecha:
                  type: string
                  format: date-time
                  example: 2026-05-01T10:00:00.000Z
      responses:
        "201":
          description: Evento creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Evento creado
                  eventos:
                    type: array
                    items:
                      $ref: "#/components/schemas/Evento"
        "400":
          description: Body no paso la validacion de Zod
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Datos invalidos
                  eventos:
                    nullable: true
                    example: null
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "500":
          $ref: "#/components/responses/ErrorServidor"

  /ausencias:
    get:
      tags: [Ausencias]
      summary: Obtener ausencias de hoy
      description: Retorna todas las ausencias del dia actual, ordenadas por fecha descendente.
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: OK o sin resultados
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: OK
                  meta:
                    $ref: "#/components/schemas/Meta"
                  ausencias:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ausencia"
              examples:
                con_resultados:
                  summary: Con ausencias
                  value:
                    mensaje: OK
                    meta:
                      total: 3
                      limit: 10
                      offset: 0
                    ausencias:
                      - id: 1
                        materia: Matematica
                        creado: "2026-04-11T08:00:00.000Z"
                        docente:
                          name: Juan Perez
                sin_resultados:
                  summary: Sin ausencias
                  value:
                    mensaje: No hay ausencias disponibles
                    meta:
                      total: 0
                      limit: 10
                      offset: 0
                    ausencias: []
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Error en el servidor
                  meta:
                    $ref: "#/components/schemas/Meta"
                  ausencias:
                    nullable: true
                    example: null
              example:
                mensaje: Error en el servidor
                meta:
                  total: 0
                  limit: null
                  offset: 0
                ausencias: null

    post:
      tags: [Ausencias]
      summary: Crear ausencia
      description: Middleware `requierePermiso("ausencias", "crear")`. Requiere sesion activa.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [materia, docenteId]
              properties:
                materia:
                  type: string
                  example: Matematica
                docenteId:
                  type: string
                  example: clxk2j3r40000abc123
      responses:
        "201":
          description: Ausencia creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Ausencia creada
                  ausencias:
                    type: array
                    items:
                      $ref: "#/components/schemas/AusenciaCreada"
        "400":
          description: Body no paso la validacion de Zod
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Datos invalidos
                  ausencias:
                    nullable: true
                    example: null
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "500":
          $ref: "#/components/responses/ErrorServidor"

  /noticias:
    get:
      tags: [Noticias]
      summary: Obtener noticias recientes
      description: Retorna todas las noticias publicadas en las ultimas 24 horas, ordenadas por fecha descendente.
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: OK o sin resultados
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: OK
                  meta:
                    $ref: "#/components/schemas/Meta"
                  noticias:
                    type: array
                    items:
                      $ref: "#/components/schemas/Noticia"
              examples:
                con_resultados:
                  summary: Con noticias
                  value:
                    mensaje: OK
                    meta:
                      total: 12
                      limit: 10
                      offset: 0
                    noticias:
                      - id: 1
                        titulo: Nueva actualizacion del sistema
                        descripcion: Se implementaron mejoras en el rendimiento
                        publicado: "2026-04-11T08:00:00.000Z"
                        recursos:
                          - id: 1
                            resource: 11-4-2026_10-30-00-123-ab3f.jpg
                            fecha: "2026-04-11T10:30:00.000Z"
                            es_noticia: true
                            noticiaId: 1
                            userId: clxk2j3r40000abc123
                sin_resultados:
                  summary: Sin noticias
                  value:
                    mensaje: No hay noticias disponibles
                    meta:
                      total: 0
                      limit: 10
                      offset: 0
                    noticias: []
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Error en el servidor
                  meta:
                    $ref: "#/components/schemas/Meta"
                  noticias:
                    nullable: true
                    example: null
              example:
                mensaje: Error en el servidor
                meta:
                  total: 0
                  limit: null
                  offset: 0
                noticias: null

    post:
      tags: [Noticias]
      summary: Crear noticia
      description: Middleware `requierePermiso("noticias", "crear")`. Requiere sesion activa. Acepta hasta 5 archivos en el campo `recursos` (jpg, jpeg, png, webp, pdf — max 10MB por archivo).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [titulo, descripcion]
              properties:
                titulo:
                  type: string
                  example: Nueva actualizacion del sistema
                descripcion:
                  type: string
                  example: Se implementaron mejoras en el rendimiento
                recursos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Noticia creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Noticia creada
                  noticias:
                    type: array
                    items:
                      $ref: "#/components/schemas/Noticia"
        "400":
          description: Body no paso la validacion de Zod
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Datos invalidos
                  noticias:
                    nullable: true
                    example: null
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/SinPermisos"
        "500":
          $ref: "#/components/responses/ErrorServidor"

  /estadisticas:
    get:
      tags: [Estadisticas]
      summary: Obtener estadisticas globales
      description: Retorna el conteo total de eventos, noticias, ausencias, usuarios y sesiones como array de objetos con clave dinamica. Middleware `requiereAuth`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: OK
                  estadisticas:
                    type: array
                    items:
                      $ref: "#/components/schemas/EstadisticaItem"
              example:
                mensaje: OK
                estadisticas:
                  - eventos: 5
                  - noticias: 12
                  - ausencias: 3
                  - usuarios: 20
                  - sesiones: 8
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Error en el servidor
                  estadisticas:
                    nullable: true
                    example: null