# apps/backend/src/routes/Ausencias/Ausencias.route.ts
import { Request, Response, Router } from "express";

import { prisma } from "@/configuracion/Prisma";
import { requiereAuth, requierePermiso } from "@/middlewares/Auth.middleware";

import { AusenciaSchema } from "./Ausencias.scheme";
import { parsePagination } from "@/Helpers/Args";

/**
 *
 * Instancia
 * 
 */
const API = Router();

/**
 * model Ausencia {
 *   id      Int      @id @default(autoincrement())
 *   materia String
 *   creado  DateTime @default(now())
 *
 *   // Docente que falta
 *   docente   User   @relation(name: "AusenciaDocente", fields: [docenteId], references: [id])
 *   docenteId String
 *
 *   // Usuario que publica
 *   publicador   User   @relation(name: "AusenciaPublicador", fields: [publicadorId], references: [id])
 *   publicadorId String
 * }
 */


API.get("/", async (req: Request, res: Response) => {
    const { limit, offset } = parsePagination(req.query)

    try {
        const [resultados, total] = await prisma.$transaction([
            prisma.ausencia.findMany({
                where: { creado: { gte: new Date(new Date().setHours(0, 0, 0, 0)) } },
                orderBy: { creado: "desc" },
                select: {
                    creado: true,
                    materia: true,
                    id: true,
                    docente: { select: { name: true } }
                },
                take: limit,
                skip: offset,
            }),
            prisma.ausencia.count({
                where: { creado: { gte: new Date(new Date().setHours(0, 0, 0, 0)) } }
            })
        ])

        const meta = { total, limit, offset }

        if (resultados.length === 0) {
            return res.status(200).json({ mensaje: "No hay ausencias disponibles", meta, ausencias: [] })
        }

        return res.status(200).json({ mensaje: "OK", meta, ausencias: resultados })

    } catch (error) {
        return res.status(500).json({ mensaje: "Error en el servidor", meta: { total: 0, limit, offset }, ausencias: null })
    }
})

API.post("/", requierePermiso("ausencias", "crear"), async (req: Request, res: Response) => {

    const resp = AusenciaSchema.safeParse(req.body)

    if (!resp.success) {
        return res.status(400).json({ mensaje: "Datos invalidos", ausencias: null })
    }

    try {
        const response = await prisma.ausencia.create({
            data: {
                ...resp.data,
                publicadorId: req.user!.id,
            },
        })

        return res.status(201).json({ mensaje: "Ausencia creada", ausencias: [response] })

    } catch (error) {
        return res.status(500).json({ mensaje: "Error en el servidor", ausencias: null })
    }
})

/**
 *
 * Export
 * 
 */
export { API as ApiAusencias }