# apps/backend/src/routes/Noticias/Noticias.route.ts
import { Request, Response, Router } from "express";

import { prisma } from "@/configuracion/Prisma";
import { Archivos } from "@/middlewares/Archivos.middleware";
import { requiereAuth, requierePermiso } from "@/middlewares/Auth.middleware";

import { NoticiaSchema } from "./Noticias.scheme";
import { parsePagination } from "@/Helpers/Args";

/**
 *
 * Instancia
 * 
 */
const API = Router();


API.get("/", async (req: Request, res: Response) => {
    const { limit, offset } = parsePagination(req.query)

    try {
        const [resultados, total] = await prisma.$transaction([
            prisma.noticia.findMany({
                where: { publicado: { gte: new Date(Date.now() - 24 * 60 * 60 * 1000) } },
                orderBy: { publicado: "desc" },
                include: { recursos: true },
                omit: { userId: true },
                take: limit,
                skip: offset
            }),
            prisma.noticia.count({
                where: { publicado: { gte: new Date(Date.now() - 24 * 60 * 60 * 1000) } }
            })
        ])

        const meta = { total, limit, offset }

        if (resultados.length === 0) {
            return res.status(200).json({ mensaje: "No hay noticias disponibles", meta, noticias: [] })
        }

        return res.status(200).json({ mensaje: "OK", meta, noticias: resultados })

    } catch (error) {
        return res.status(500).json({ mensaje: "Error en el servidor", meta: { total: 0, limit, offset }, noticias: null })
    }
})

API.post("/",

    /** Middlewares */
    requierePermiso("noticias", "crear"),

    Archivos({
        formatos: [".jpg", ".jpeg", ".png", ".webp", ".pdf"],
        maxFiles: 5,
        maxSizeFile: 10 * 1024 * 1024, // 10mb
    }).array("recursos"),

    // Handle
    async (req: Request, res: Response) => {

        const resp = NoticiaSchema.safeParse(req.body)

        if (!resp.success) {
            return res.status(400).json({ mensaje: "Datos invalidos", noticias: null })
        }

        try {
            const response = await prisma.noticia.create({
                data: {
                    ...resp.data,
                    userId: req.user!.id,
                    recursos: {
                        create: (req.files as Express.Multer.File[] ?? []).map(file => ({
                            resource: file.filename,
                            es_noticia: true,
                            userId: req.user!.id,
                        }))
                    }
                },
                include: {
                    recursos: true
                },
                omit: {
                    userId: true
                }
            })

            return res.status(201).json({ mensaje: "Noticia creada", noticias: [response] })

        } catch (error) {
            return res.status(500).json({ mensaje: "Error en el servidor", noticias: null })
        }
    })


/**
 *
 * Export
 * 
 */
export { API as ApiNoticias }